Fix generated code

+++ demo/src/main/java/com/example/demo/bl/DoctorBl.java
@@ -248,12 +248,12 @@ public class DoctorBl {
     }
 
     private String generateCode(int id){
-        String code = String.valueOf(id)+""+(String.valueOf(Math.random() * (9999 - 999)) + 999);
+        String code = String.valueOf(id)+""+(String.valueOf((int) (Math.random()*(9999-999)+999)));
         System.out.println("Generated Code: ->"+code);
         return code;
     }
     private String generateCodePass(int id){
-        String code = String.valueOf(id)+"-"+(String.valueOf(Math.random() * (9999 - 999)) + 999);
+        String code = String.valueOf(id)+"-"+(String.valueOf((int) (Math.random()*(9999-999)+999)));
         System.out.println("Generated Code: ->"+code);
         return code;
     }
